+++ deltafi-core/src/main/java/org/deltafi/core/snapshot/SnapshotRestoreOrder.java
@@ -24,6 +24,8 @@ public class SnapshotRestoreOrder {
     public static final int PROPERTIES_ORDER = Ordered.HIGHEST_PRECEDENCE;
     public static final int DELETE_POLICY_ORDER = Ordered.HIGHEST_PRECEDENCE + 10;
     public static final int FLOW_ASSIGNMENT_ORDER = Ordered.HIGHEST_PRECEDENCE + 10;
+    public static final int PLUGIN_CUSTOMIZATION_CONFIG_ORDER = Ordered.HIGHEST_PRECEDENCE + 15;
+    public static final int PLUGIN_IMAGE_REPO_ORDER = Ordered.HIGHEST_PRECEDENCE + 15;
     public static final int PLUGIN_REGISTRY_ORDER = Ordered.HIGHEST_PRECEDENCE + 20;
     public static final int PLUGIN_VARIABLE_ORDER = Ordered.HIGHEST_PRECEDENCE + 30;
     public static final int FLOW_PLAN_ORDER = Ordered.HIGHEST_PRECEDENCE + 40;